carrier-messenger-app/Aldridge-files/requirements.txt
# Python dependencies required to run the Flask chat application
flask
redis
python-dotenv
gunicorn
mysql-connector-python
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
python-dateutil
Flask-APScheduler
